% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_rdstemplate}
\alias{create_rdstemplate}
\title{Creates New Project}
\usage{
create_rdstemplate(path, use_packrat, install_default_packages, project_name,
  ...)
}
\arguments{
\item{path}{path to the new project}

\item{use_packrat}{TRUE/FALSE should we use packrat with the new project or not}

\item{install_default_packages}{should we install default packages from inst/default_packages.txt?
Works only if \code{use_packrat} is TRUE}

\item{project_name}{Name of creating project. It can differ from path.}

\item{...}{Left for compatibility}
}
\description{
Creates New Project
}
